3-3-2023

+++ LMS.Data/Models/BaseModel.cs
@@ -6,9 +6,9 @@ namespace LMS.Data.Models
 {
     public class BaseModel
     {
-        public long? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
         public DateTime? CreatedDate { get; set; }
-        public long? UpdatedBy { get; set; }
+        public int? UpdatedBy { get; set; }
         public DateTime? UpdatedDate { get; set; }
         public bool IsDelete { get; set; }
     }
